+++ R/variant_standard.R
@@ -191,7 +191,7 @@ get_group_level_standard <- function(parameters, s){
 }
 
 fill_samples_standard <- function(samples, group_level, proposals, j = 1, n_pars){
-  samples$a_half[, j] <- group_level$a_half
+  emc_set_last_slice(samples$a_half, j, as.numeric(group_level$a_half))
   samples$last_theta_var_inv <- group_level$tvinv
   samples <- fill_samples_base(samples, group_level, proposals, j = j, n_pars)
   return(samples)